Modularize methods that output doc pages and examples
